+++ lib/models/dto/meal_dto.g.dart
@@ -8,19 +8,21 @@ part of 'meal_dto.dart';
 
 MealDto _$MealDtoFromJson(Map<String, dynamic> json) => MealDto(
       mealId: json['mealId'] as String,
-      count: (json['count'] as num).toInt(),
+      count: (json['count'] as num?)?.toInt(),
+      packagesCount: (json['packagesCount'] as num?)?.toInt(),
       preparedAt: _$JsonConverterFromJson<Timestamp, DateTime>(
           json['preparedAt'], const TimestampConverter().fromJson),
       consumeBy: _$JsonConverterFromJson<Timestamp, DateTime>(
           json['consumeBy'], const TimestampConverter().fromJson),
-      foodBoxId: json['foodBoxId'] as String,
-      foodBoxCount: (json['foodBoxCount'] as num).toInt(),
+      foodBoxId: json['foodBoxId'] as String?,
+      foodBoxCount: (json['foodBoxCount'] as num?)?.toInt(),
       foodTemperature: (json['foodTemperature'] as num?)?.toInt(),
     );
 
 Map<String, dynamic> _$MealDtoToJson(MealDto instance) => <String, dynamic>{
       'mealId': instance.mealId,
       'count': instance.count,
+      'packagesCount': instance.packagesCount,
       'preparedAt': _$JsonConverterToJson<Timestamp, DateTime>(
           instance.preparedAt, const TimestampConverter().toJson),
       'consumeBy': _$JsonConverterToJson<Timestamp, DateTime>(