+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This is a changelog for **ZachranObed** application.
 
 ### Added
 - **ZOB-278** Add web navigation drawer.
+- **ZOB-308** Add packaged food category.
 
 ### Fixed